requirements.txt
certifi==2018.4.16
future==0.16.0
import-string==0.1.0
python-telegram-bot==10.1.0
six==1.11.0
